Move discovery to extra file
